requirements.txt
diffusers
transformers
accelerate
torch
xformers
gradio
Pillow
qrcode
filelock